fix：完善引擎功能；适配mac/win平台接入

+++ engine/core/gfx/base/gfx-buffer-instance.cpp
@@ -6,7 +6,7 @@
 GfxBufferInstance::GfxBufferInstance()
 {
 }
-GfxBuffer *GfxBufferInstance::getBuffer(int size)
+GfxBuffer *GfxBufferInstance::getBuffer(size_t size)
 {
     // 检查是否存在该大小的池
     if (this->pools.find(size) == this->pools.end())
